Move tests around to better model use-case.
The with(){} was being evaluated on each run, rather than just once and have the function defined within it.  In the case under consideration, a with(){} will be evaluated on each module load, and then the module functions will be defined within it.

Thus, the purpose of the test is not to measure how much time interpreting a with(){} adds to execution, but rather the effect on the resulting code of having the extra object in the lookup chain.